requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.7.4
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
websockets==8.1
yarl==1.4.2
